pyproject.toml
[tool.poetry]
name = "fastapi_tut"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
Jinja2 = "^3.0.3"
aiofiles = "^0.8.0"
SQLAlchemy = "^1.4.29"
psycopg2 = "^2.9.3"
python-dotenv = "^0.19.2"
pydantic = {extras = ["email"], version = "^1.9.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
Faker = "^11.3.0"
alembic = "^1.7.5"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"
pytest = "^6.2.5"
requests = "^2.27.1"
httpx = "^0.22.0"
trio = "^0.19.0"
fastapi-jwt-auth = "^0.5.0"
sqlmodel = "^0.0.6"
factory-boy = "^3.2.1"
marshmallow-sqlalchemy = "^0.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
#log_cli = "True"
#log_cli_level = "INFO"
log_cli_format = "%(asctime)s %(levelname)s %(message)s"
log_file = "logs/pytest-logs.txt"
log_file_level = "INFO"
log_file_format = "%(asctime)s %(levelname)s %(message)s"
log_level = "INFO"

